+++ browser-switch/src/main/java/com/braintreepayments/browserswitch/BrowserSwitchResult.java
@@ -1,0 +1,28 @@
+package com.braintreepayments.browserswitch;
+
+public class BrowserSwitchResult {
+
+    private int status;
+    private String errorMessage;
+
+    public static final int STATUS_OK = 1;
+    public static final int STATUS_CANCELLED = 2;
+    public static final int STATUS_ERROR = 3;
+
+    BrowserSwitchResult(int status) {
+        this(status, null);
+    }
+
+    BrowserSwitchResult(int status, String errorMessage) {
+        this.status = status;
+        this.errorMessage = errorMessage;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+}