+++ demo/src/main/java/com/braintreepayments/browserswitch/demo/DemoFragment.java
@@ -8,26 +8,27 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.braintreepayments.browserswitch.BrowserSwitchFragment;
-
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.braintreepayments.browserswitch.BrowserSwitchFragment;
+import com.braintreepayments.browserswitch.BrowserSwitchResult;
 
 public class DemoFragment extends BrowserSwitchFragment implements View.OnClickListener {
 
-    private Button mBrowserSwitchButton;
     private TextView mResult;
     private TextView mReturnUrl;
 
     @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.demo_fragment, null);
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.demo_fragment, container, false);
 
-        mBrowserSwitchButton = (Button) view.findViewById(R.id.browser_switch);
-        mBrowserSwitchButton.setOnClickListener(this);
+        Button browserSwitchButton = view.findViewById(R.id.browser_switch);
+        browserSwitchButton.setOnClickListener(this);
 
-        mResult = (TextView) view.findViewById(R.id.result);
-        mReturnUrl = (TextView) view.findViewById(R.id.return_url);
+        mResult = view.findViewById(R.id.result);
+        mReturnUrl = view.findViewById(R.id.return_url);
 
         return view;
     }
@@ -41,7 +42,25 @@ public class DemoFragment extends BrowserSwitchFragment implements View.OnClickL
 
     @Override
     public void onBrowserSwitchResult(int requestCode, BrowserSwitchResult result, @Nullable Uri returnUri) {
-        mResult.setText("Result: " + result.name());
-        mReturnUrl.setText("Return url: " + returnUri);
+        String resultText = null;
+        String returnUrl = "";
+
+        int statusCode = result.getStatus();
+        switch (statusCode) {
+            case BrowserSwitchResult.STATUS_OK:
+                resultText = "Browser Switch Successful";
+                if (returnUri != null) {
+                    returnUrl = "Uri: " + returnUri.toString();
+                }
+                break;
+            case BrowserSwitchResult.STATUS_ERROR:
+                resultText = "Browser Switch Error: " + result.getErrorMessage();
+                break;
+            case BrowserSwitchResult.STATUS_CANCELLED:
+                resultText = "Browser Switch Cancelled by User";
+                break;
+        }
+        mResult.setText(resultText);
+        mReturnUrl.setText(returnUrl);
     }
 }