+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # browser-switch-android Release Notes
 
+## unreleased
+
+* Create BrowserSwitchClient to allow browser switch behavior through composition as well as inheritance.
+* Update minSdkVersion from 15 to 21 (thanks! @calvarez-ov)
+
 ## 0.2.0
 
 * Using androidx 1.0.0